# openapi.yml
openapi: 3.0.0
info:
  title: INT AI API
  version: 1.0.0
  description: Backend API for INT AI platform
servers:
  - url: http://localhost:3001
    description: Development server
paths:
...
  /api/connections/oauth/authorize/{provider}:
    get:
      summary: Initiate OAuth flow
      security:
        - bearerAuth: []
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
        - name: workspaceId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OAuth authorization URL
  /api/connections/oauth/callback/{provider}:
    get:
      summary: OAuth callback
      security:
        - bearerAuth: []
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
        - name: code
          in: query
          required: true
          schema:
            type: string
        - name: state
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Connection created
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT